% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\encoding{UTF-8}
\name{translate_queries}
\alias{translate_queries}
\title{Replace special characters in queries}
\usage{
translate_queries(x)
}
\arguments{
\item{x}{a string to tranlate}
}
\description{
Translate scandinavian characters to code for query.
}
\examples{
 translate_queries("V\u00e4est\u00f6 V\u00c4EST\u00d6")
}
