% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uljas_user_functions.R
\name{uljas_data}
\alias{uljas_data}
\title{Get data from Uljas api}
\usage{
uljas_data(classifiers, ifile, lang = "en", naming = "label", ...)
}
\arguments{
\item{classifiers}{a list of classes with values of levels to get.}

\item{ifile}{a name of the statistics file with relative data directory path.
Get from \code{\link{uljas_stats}}}

\item{lang}{a language code. Available en, fi, se.}

\item{naming}{whether to use (longer) labels or (shorter) ids.}

\item{...}{additional parameters for a query. Passed to \code{\link{uljas_api}}.}
}
\value{
a tibble with a column for each classifier and a values column.
}
\description{
\code{uljas_data} returns the data for class value combinations from a statistics
(specified with ifile parameter).
}
\details{
The names of \code{classifiers} have to be classification names of the
statistics file, see \code{\link{uljas_dims}} and \code{\link{uljas_class}}.
A query with an unknown name or an unknown code fails with a status code 500.

A classifier value may also be one of the api keywords \code{"=ALL"},
\code{"=FIRST"} and \code{"=LAST"}. Note that the keywords follow the order
of the classification, which for time periods is the newest first, and that
the code of a total is an empty string.
}
\examples{
\donttest{
  sitc_query <- list(`Classification of Products SITC1` = c("0" , "1"),
                     `Time period` = "=ALL", Flow = 1, Country = "AT", Indicators = "V1")
  sitc_data <- uljas_data(ifile = "/DATABASE/01 ULKOMAANKAUPPATILASTOT/02 SITC/ULJAS_SITC",
                          classifiers = sitc_query)
}
}
