% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{uljas_query_size}
\alias{uljas_query_size}
\title{Number of cells a query asks for}
\usage{
uljas_query_size(classifiers, ifile, lang = "en")
}
\arguments{
\item{classifiers}{a list of classes with values of levels to get.}

\item{ifile}{a name of the statistics file with relative data directory path.
Get from \code{\link{uljas_stats}}}

\item{lang}{a language code. Available en, fi, se.}
}
\value{
a number, NA if the size of a classification is not known.
}
\description{
\code{uljas_query_size} returns the number of cells, i.e. the number of
classifier value combinations, that a query asks for. An individual request
is limited to 50 000 cells and \code{\link{uljas_data}} divides a larger
query into several requests.
}
\details{
The sizes of the classifications are asked for with
\code{\link{uljas_dims}}, if the query uses the keyword \code{"=ALL"}.
}
\examples{
\donttest{
  uljas_query_size(
    ifile = "/DATABASE/01 ULKOMAANKAUPPATILASTOT/02 SITC/ULJAS_SITC",
    classifiers = list(`Classification of Products SITC5` = "=ALL",
                       `Time period` = "=ALL", Flow = 1, Country = "AT",
                       Indicators = "V1"))
}
}
