% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uljas_user_functions.R
\name{uljas_stats}
\alias{uljas_stats}
\title{Get statistics from the Uljas database.}
\usage{
uljas_stats(lang = "en")
}
\arguments{
\item{lang}{a language code. Available en, fi, se.}
}
\value{
a data.frame with the columns ifile, title and utime.
}
\description{
Returns a data.frame of statistics in the Uljas database.
}
\examples{
\donttest{
  stats <- uljas_stats("fi")
}
}
