% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_queries}
\alias{translate_queries}
\title{Replace special characters in queries}
\usage{
translate_queries(x)
}
\arguments{
\item{x}{a string to tranlate}
}
\description{
Translate scandinavian characters to code for query.
}
\examples{
 translate_queries("Väestö VÄESTÖ")
}
