% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uljas_api.R
\name{uljas_api}
\alias{uljas_api}
\title{Main api-function for R client for the Finnish Customs (Tulli) database Uljas}
\usage{
uljas_api(lang = "en", atype, konv, ..., query_list = NULL, naming = "label")
}
\arguments{
\item{lang}{A language code. Available en, fi, se.}

\item{atype}{A type of query. One of stats, dims, class or data.}

\item{konv}{A output format. One of json or json-stat.}

\item{...}{Additional parameters for a query.}

\item{query_list}{Additional parameters for a query as a list.}

\item{naming}{whether to use (longer) labels or (shorter) ids.}
}
\value{
A list with uljas api type class.
}
\description{
Use user functions \code{\link{uljas_stats}}, \code{\link{uljas_dims}},
\code{\link{uljas_class}} and \code{\link{uljas_data}}.
See more from the Uljas api page
\url{https://tilastot.tulli.fi/en/uljas-statistical-database/uljas-api}.
}
\details{
A query is sent as a get request, or as a post request if it is longer than
the 2048 bytes that the server accepts in an url.
}
