% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uljas_api.R
\name{uljas_api}
\alias{uljas_api}
\title{Main api-function for R client for the Finnish Customs (Tulli) database Uljas}
\usage{
uljas_api(lang = "en", atype, konv, ..., query_list = NULL)
}
\arguments{
\item{lang}{A language code. Available en, fi, se.}

\item{atype}{A type of query.}

\item{konv}{A output format}

\item{...}{Additional parameters for a query.}

\item{query_list}{Additional parameters for a query as a list.}
}
\value{
A list with uljas api type class.
}
\description{
Use user functions \code{\link{uljas_stats}}, \code{\link{uljas_dims}},
\code{\link{uljas_class}} and \code{\link{uljas_data}}.
See more from Uljas api page https://tulli.fi/en/statistics/uljas-api
}
